Suppression des views inutiles

+++ src/Router.php
@@ -17,6 +17,7 @@ class Router
     }
 
     public function main() {
+
         session_start();
 
         $feedback = key_exists('feedback', $_SESSION) ? $_SESSION['feedback'] : '';
@@ -47,7 +48,6 @@ class Router
 
         }
 
-
         try{
             switch ($action) {
                 case "voir":
@@ -68,37 +68,51 @@ class Router
                     $ctrl->allUsersWithSeriesPage();
                     break;
                 case "creerSerie" :
-                    $ctrl->newSerie();
+                    if ( key_exists('pseudo', $_SESSION))
+                        $ctrl->newSerie();
                     break;
+
                 case "sauverNouvelleSerie" :
-                    $ctrl->saveNewSerie($_POST);
+                    if ( key_exists('pseudo', $_SESSION))
+                        $ctrl->saveNewSerie($_POST);
                     break;
                 case "creerManga" :
-                    $ctrl->newManga(null);
+                    if ( key_exists('pseudo', $_SESSION))
+                        $ctrl->newManga(null);
                     break;
                 case "sauverNouveauManga" :
-                    $ctrl->saveNewManga($_POST);
+                    if ( key_exists('pseudo', $_SESSION))
+                        $ctrl->saveNewManga($_POST);
                     break;
                 case "supprimer" :
-                    if ($userPseudo === null || $serieId === null || $tomeId === null) {
-                        $view->makeUnknownActionPage();
-                    } else {
-                        $ctrl->deleteManga($userPseudo, $serieId, $tomeId);
+                    if ( key_exists('pseudo', $_SESSION)){
+
+                        if ($userPseudo === null || $serieId === null || $tomeId === null) {
+                            $view->makeUnknownActionPage();
+                        } else {
+                            $ctrl->deleteManga($userPseudo, $serieId, $tomeId);
+                        }
                     }
                     break;
                 case "sauverNouveauCompte" :
-
-                    $ctrl->saveNewCompte($_POST);
+                    if ( !key_exists('pseudo', $_SESSION))
+                        $ctrl->saveNewCompte($_POST);
                     break;
                 case "connexion" :
-                    $ctrl->connexion();
+                    if ( !key_exists('pseudo', $_SESSION))
+                        $ctrl->connexion();
+                    break;
+                case "deconnecter" :
+                    if ( key_exists('pseudo', $_SESSION))
+                        $ctrl->deconnexion();
                     break;
                 case "sauverConnexion" :
-
-                    $ctrl->saveConn($_POST);
+                    if ( !key_exists('pseudo', $_SESSION))
+                        $ctrl->saveConn($_POST);
                     break;
                 case "creerCompte" :
-                    $ctrl->newCompte();
+                    if ( !key_exists('pseudo', $_SESSION))
+                        $ctrl->newCompte();
                     break;
             }
         }catch (Exception $e) {
@@ -163,6 +177,10 @@ class Router
 
     public function inscriptionPage(){
         return ".?action=creerCompte";
+    }
+
+    public function deconnecter(){
+        return ".?action=deconnecter";
     }
 
     public function POSTredirect($url, $feedback) {
